JCR-749 Add mysql ddl for clustering (DatabaseJournal) JCR-756 Concurrent add/remove child node operations in a cluster may corrupt repository JCR-757 Allow multiple producers to feed/consume journal
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@509624 13f79535-47bb-0310-9956-ffa450edef68

+++ jackrabbit-core/src/main/java/org/apache/jackrabbit/core/config/ClusterConfig.java
@@ -30,7 +30,7 @@ public class ClusterConfig {
     /**
      * Sync delay.
      */
-    private final int syncDelay;
+    private final long syncDelay;
 
     /**
      * Journal configuration.
@@ -41,9 +41,10 @@ public class ClusterConfig {
      * Creates a new cluster configuration.
      *
      * @param id custom cluster node id
+     * @param syncDelay syncDelay, in milliseconds
      * @param jc journal configuration
      */
-    public ClusterConfig(String id, int syncDelay, JournalConfig jc) {
+    public ClusterConfig(String id, long syncDelay, JournalConfig jc) {
         this.id = id;
         this.syncDelay = syncDelay;
         this.jc = jc;
@@ -63,7 +64,7 @@ public class ClusterConfig {
      *
      * @return syncDelay
      */
-    public int getSyncDelay() {
+    public long getSyncDelay() {
         return syncDelay;
     }
